server/requirements.txt
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==3.20.1
Werkzeug==2.1.2
zipp==3.8.0
